+++ ACP3/Modules/ACP3/Articles/Resources/config/services.yml
@@ -2,10 +2,9 @@ imports:
   - { resource: components/cache.yml }
   - { resource: components/controllers.yml }
   - { resource: components/events.yml }
+  - { resource: components/extension.yml }
   - { resource: components/installer.yml }
   - { resource: components/models.yml }
-  - { resource: components/search.yml }
-  - { resource: components/seo.yml }
   - { resource: components/validation.yml }
 
 services:
