+++ ACP3/Modules/ACP3/Files/Extension/FeedAvailabilityExtension.php
@@ -4,17 +4,17 @@
  * See the LICENCE file at the top-level module directory for licencing details.
  */
 
-namespace ACP3\Modules\ACP3\Files\Feeds;
+namespace ACP3\Modules\ACP3\Files\Extension;
 
 
 use ACP3\Core\Date;
 use ACP3\Core\Helpers\StringFormatter;
 use ACP3\Core\Router\RouterInterface;
-use ACP3\Modules\ACP3\Feeds\Utility\FeedAvailabilityInterface;
+use ACP3\Modules\ACP3\Feeds\Extension\FeedAvailabilityExtensionInterface;
 use ACP3\Modules\ACP3\Files\Installer\Schema;
 use ACP3\Modules\ACP3\Files\Model\Repository\FilesRepository;
 
-class FeedAvailability implements FeedAvailabilityInterface
+class FeedAvailabilityExtension implements FeedAvailabilityExtensionInterface
 {
     /**
      * @var \ACP3\Core\Date
