+++ ACP3/Modules/ACP3/Files/Extension/SearchAvailabilityExtension.php
@@ -4,16 +4,16 @@
  * See the LICENCE file at the top-level module directory for licencing details.
  */
 
-namespace ACP3\Modules\ACP3\Articles\Search;
+namespace ACP3\Modules\ACP3\Files\Extension;
 
 
 use ACP3\Core\Date;
 use ACP3\Core\Router\RouterInterface;
-use ACP3\Modules\ACP3\Articles\Installer\Schema;
-use ACP3\Modules\ACP3\Articles\Model\Repository\ArticleRepository;
-use ACP3\Modules\ACP3\Search\Utility\SearchAvailabilityInterface;
+use ACP3\Modules\ACP3\Files\Installer\Schema;
+use ACP3\Modules\ACP3\Files\Model\Repository\FilesRepository;
+use ACP3\Modules\ACP3\Search\Extension\SearchAvailabilityExtensionInterface;
 
-class SearchAvailability implements SearchAvailabilityInterface
+class SearchAvailabilityExtension implements SearchAvailabilityExtensionInterface
 {
     /**
      * @var \ACP3\Core\Date
@@ -24,24 +24,24 @@ class SearchAvailability implements SearchAvailabilityInterface
      */
     private $router;
     /**
-     * @var \ACP3\Modules\ACP3\Articles\Model\Repository\ArticleRepository
+     * @var \ACP3\Modules\ACP3\Files\Model\Repository\FilesRepository
      */
-    private $articleRepository;
+    private $filesRepository;
 
     /**
      * SearchAvailability constructor.
      * @param Date $date
      * @param RouterInterface $router
-     * @param ArticleRepository $articleRepository
+     * @param FilesRepository $filesRepository
      */
     public function __construct(
         Date $date,
         RouterInterface $router,
-        ArticleRepository $articleRepository
+        FilesRepository $filesRepository
     ) {
         $this->date = $date;
         $this->router = $router;
-        $this->articleRepository = $articleRepository;
+        $this->filesRepository = $filesRepository;
     }
 
     /**
@@ -62,7 +62,7 @@ class SearchAvailability implements SearchAvailabilityInterface
     {
         $fields = $this->mapSearchAreasToFields($areas);
 
-        $results = $this->articleRepository->getAllSearchResults(
+        $results = $this->filesRepository->getAllSearchResults(
             $fields,
             $searchTerm,
             $sortDirection,
@@ -71,7 +71,7 @@ class SearchAvailability implements SearchAvailabilityInterface
         $cResults = count($results);
 
         for ($i = 0; $i < $cResults; ++$i) {
-            $results[$i]['hyperlink'] = $this->router->route('articles/index/details/id_' . $results[$i]['id']);
+            $results[$i]['hyperlink'] = $this->router->route('files/index/details/id_' . $results[$i]['id']);
         }
 
         return $results;
@@ -86,11 +86,11 @@ class SearchAvailability implements SearchAvailabilityInterface
     {
         switch ($areas) {
             case 'title':
-                return 'title';
+                return 'title, file';
             case 'content':
                 return 'text';
             default:
-                return 'title, text';
+                return 'title, file, text';
         }
     }
 }
