--- ACP3/Modules/ACP3/Files/Resources/config/components/feeds.yml
@@ -1,10 +1,0 @@
-services:
-  files.feeds.feed_availability:
-    class: ACP3\Modules\ACP3\Files\Feeds\FeedAvailability
-    arguments:
-      - '@core.date'
-      - '@core.router'
-      - '@core.helpers.string_formatter'
-      - '@files.model.filesrepository'
-    tags:
-      - { name: feeds.feed_availability }
