--- ACP3/Modules/ACP3/Files/Resources/config/components/search.yml
@@ -1,9 +1,0 @@
-services:
-  files.search.search_availability:
-    class: ACP3\Modules\ACP3\Files\Search\SearchAvailability
-    arguments:
-      - '@core.date'
-      - '@core.router'
-      - '@files.model.filesrepository'
-    tags:
-      - { name: search.search_availability }
