--- ACP3/Modules/ACP3/News/Resources/config/components/feeds.yml
@@ -1,10 +1,0 @@
-services:
-  news.feeds.feed_availability:
-    class: ACP3\Modules\ACP3\News\Feeds\FeedAvailability
-    arguments:
-      - '@core.date'
-      - '@core.router'
-      - '@core.helpers.string_formatter'
-      - '@news.model.newsrepository'
-    tags:
-      - { name: feeds.feed_availability }
