--- ACP3/Modules/ACP3/News/Resources/config/components/search.yml
@@ -1,9 +1,0 @@
-services:
-  news.search.search_availability:
-    class: ACP3\Modules\ACP3\News\Search\SearchAvailability
-    arguments:
-      - '@core.date'
-      - '@core.router'
-      - '@news.model.newsrepository'
-    tags:
-      - { name: search.search_availability }
